refactor: move util functions from basic to utils (refs 55)
